backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
python-dotenv==1.0.0
qdrant-client==1.7.0
fasttext==0.9.3
pandas==2.2.2
openpyxl==3.1.5
beautifulsoup4==4.12.3
tqdm==4.66.5
numpy==2.0.1